+++ src/lib/UI/domain-components/AddDemoCoffeeBeans_PageBlock.svelte
